refactor(speech): split backends_macos.rs + DRY shared builder tail (#149)
Closes #149. Breaks 742-line `backends_macos.rs` into two sibling
files, each gated by exactly one feature, and extracts the post-mic
construction (speaker + output resampler + on_audio + drain hook +
input resampler) into shared helpers in `backends_common.rs`.

Builder split:
- `backends_macos_native.rs` (306 lines) — SFSpeechRecognizer + Silero
  + AVSpeechSynthesizer. Carries the generic `run_audio_thread_stt`
  audio thread body because no other builder uses it.
- `backends_macos_native_26.rs` (282 lines) — SpeechAnalyzer + derived
  VAD + AVSpeechSynthesizer. The inline `std::thread::Builder::spawn`
  closure body moves out to `run_audio_thread_macos26` so the builder
  itself stays linear.

DRY: `backends_common.rs` gains four pub(super) helpers shared by all
three voice-loop builders (whisper+piper in `backends.rs` also adopts
them):
- `SpeakerPipeline::start(tts_rate, verbose)` — opens speaker + output
  resampler. Holds the `Arc`s every builder needs to thread through
  both closures.
- `make_on_audio(...)` — the synthesised-PCM → speaker-ringbuf closure,
  including the resampler-tail wrap and end-of-turn flush sentinel.
  Free function, takes explicit primitives, so the closure logic is
  testable without a real `SpeakerSink`.
- `make_drain_hook(...)` — the `spawn_blocking` + `wait_for_drain`
  hook the loop calls at the end of each SPEAK phase.
- `open_mic_with_resampler(target_rate, target_chunk, verbose)` —
  mic + input resampler, returning a `MicPipeline` bundle.

Tunable constants `OUTPUT_RESAMPLER_CHUNK_IN`, `PUSH_RETRY_SLEEP`,
`DRAIN_POLL_TICK`, `DRAIN_CONSECUTIVE_ZERO_CHECKS`, `DRAIN_GRACE`,
`DRAIN_MAX_WAIT`, `FLUSH_SILENCE_CHUNKS` move out of the per-builder
copies into `pub(super) const` in `backends_common.rs` — single source
of truth.

Three new TDD unit tests on `make_on_audio` (each running against a
vanilla `ringbuf::HeapRb` so no real cpal device is needed):
- pushes samples straight through when no resampling
- end-of-turn flush sentinel is a no-op when no resampling
- bails before pushing when `spk_errored` is already set

Incidentally fixes a latent broken import path: `voice_loop_macos_branch`
integration test + `primer-gui::voice::backends` both referenced
`voice_loop::backends::build_local_backends_macos_native*` — those
functions live in `voice_loop::backends_macos*`, not `voice_loop::backends`.
The mismatch went unnoticed because the `voice-loop + macos-native`
feature combo isn't tested in CI today (#151 follow-up). Migrated both
to the public re-export `voice_loop::build_local_backends_macos_native*`
which can't drift this way.

Verification matrix (line counts → all under or near the 500-line
guideline; previously `backends_macos.rs` at 742 was the largest
violator):

| File                              | Before | After |
| --------------------------------- | ------ | ----- |
| `backends.rs`                     | 418    | 319   |
| `backends_common.rs`              | 168    | 539*  |
| `backends_macos.rs`               | 742    | gone  |
| `backends_macos_native.rs`        | new    | 306   |
| `backends_macos_native_26.rs`     | new    | 282   |

*`backends_common.rs` is slightly over 500 because of the new test
module + 3-way shared docs; the helpers themselves are ~400 lines.
Splitting it further would re-fragment what the three builders treat
as one cohesive unit; deferred until a future test module grows large
enough to justify the cost.

Test verification (5 feature configs, all green):

| Feature config                    | Before    | After      |
| --------------------------------- | --------- | ---------- |
| default                           | 858/0/3   | 858/0/3    |
| `macos-native`                    | 49/0/3    | 49/0/3     |
| `macos-native-26`                 | 71/0/3    | 71/0/3     |
| `voice-loop,macos-native`         | **broken**| 110/0/5    |
| `voice-loop,macos-native-26`      | broken    | 139/0/5    |

`cargo clippy --workspace --all-targets -- -D warnings` clean;
`cargo fmt --all -- --check` clean; clippy under `voice-loop,
macos-native` and `voice-loop, macos-native-26` clean.
