+++ src/crates/primer-speech/src/voice_loop/macos26_audio_buffer.rs
@@ -24,7 +24,7 @@
 //! edge) so the helper stays stateless beyond the buffer itself — same
 //! shape as the matching `raw_buf.clear()` in
 //! [`super::backends::run_audio_thread`] (the whisper path) and
-//! [`super::backends_macos::run_audio_thread_stt`] (the
+//! [`super::backends_macos_native::run_audio_thread_stt`] (the
 //! SFSpeechRecognizer path), both of which have always handled this
 //! correctly.
 //!
